+++ tailwind.config.js
@@ -13,34 +13,7 @@ module.exports = {
         extend: {
             fontFamily: {
                 sans: ['Inter var', ...defaultTheme.fontFamily.sans],
-            },
-            colors: {
-                current: 'currentColor',
-                transparent: 'transparent',
-                'blue-gray': colors.blueGray,
-                'cool-gray': colors.coolGray,
-                'pink-gray': colors.pinkGray,
-                'true-Gray': colors.trueGray,
-                'warm-Gray': colors.warmGray,
-                gray: colors.gray,
-                red: colors.red,
-                orange: colors.orange,
-                amber: colors.amber,
-                yellow: colors.yellow,
-                lime: colors.lime,
-                green: colors.green,
-                emerald: colors.emerald,
-                teal: colors.teal,
-                cyan: colors.cyan,
-                sky: colors.sky,
-                blue: colors.blue,
-                indigo: colors.indigo,
-                violet: colors.violet,
-                purple: colors.purple,
-                fuchsia: colors.fuchsia,
-                pink: colors.pink,
-                rose: colors.rose,      
-              }            
+            }                        
         },
     },